chore: add EthApiError::Other
